backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlmodel==0.0.18
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
neo4j==5.21.0
redis==5.0.4
celery==5.4.0
dspy-ai==2.4.9
pulp==2.8.0
httpx==0.27.0
playwright==1.49.0
orjson==3.10.3
pytest==8.2.2
pytest-asyncio==0.23.7
respx==0.21.1
